fix avatar

+++ src/controllers/user.controller.js
@@ -1,5 +1,4 @@
 require("dotenv").config();
-const multer = require("multer");
 const axios = require("axios");
 const fs = require("fs");
 const userService = require("../services/user.service");
@@ -21,26 +20,33 @@ const updateUserInfo = async (req, res) => {
     if (!userId) {
       res.status(500).json({
         message: "Lỗi khi cập nhật thông tin người dùng",
-        error: err.response ? err.response.data : err.message,
       });
     }
-    const imageFilePath = req.file.path;
-    const imageData = fs.readFileSync(imageFilePath);
+    let avatar;
 
-    const form = new FormData();
-    form.append("image", imageData.toString("base64"));
+    if (req.file) {
+      const imageFilePath = req.file.path;
 
-    const response = await axios.post("https://api.imgur.com/3/image", form, {
-      headers: {
-        Authorization: `Client-ID ${process.env.IMGUR_CLIENT_ID}`,
-        ...form.getHeaders(),
-      },
-      timeout: 10000,
-    });
+      const imageData = fs.readFileSync(imageFilePath);
 
-    fs.unlinkSync(imageFilePath);
+      const form = new FormData();
+      form.append("image", imageData.toString("base64"));
 
-    const avatar = response.data.data.link;
+      const response = await axios.post("https://api.imgur.com/3/image", form, {
+        headers: {
+          // eslint-disable-next-line no-undef
+          Authorization: `Client-ID ${process.env.IMGUR_CLIENT_ID}`,
+          ...form.getHeaders(),
+        },
+        timeout: 10000,
+      });
+
+      fs.unlinkSync(imageFilePath);
+      avatar = response.data.data.link;
+    } else {
+      // Nếu không có file, sử dụng trực tiếp text được gửi
+      avatar = req.body.avatar;
+    }
 
     await userService.updateUserInfo(userId, req.body, avatar);
